Added `entity.name.function` selector
Added rules to define `entity.name.function` selector, so that “Goto Symbol” is able to find definitions in ATS source files.

Note Sublime Text seems to only recognize `entity.name.function` instead of any `entity.name` as one would expect. That's why I included `val` and `var`, which may anyway more or less be seen as kind of functions, however, not type definition. I have not added `datatype`, `dataview` and so on, as I'm personally not happy to give them the `entity.name.function`, while I believe this would be nice to be able to find type definition with the “Goto Symbol” feature of Sublime Text.